structure files into separate folders
