+++ front-end/src/App.js
@@ -1,16 +1,16 @@
 import './css/App.css'
 import React, { useState, useEffect } from 'react'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
-import Home from './Home.js'
-import TermsOfService from './TermsOfService'
-import MeetTheTeam from './MeetTheTeam'
-import ScrollToTop from './ScrollToTop'
-import Profile from './Profile'
-import MyShows from './MyShows'
-import IndividualShow from './IndividualShow'
+import Home from './main/Home.js'
+import TermsOfService from './main/TermsOfService'
+import MeetTheTeam from './main/MeetTheTeam'
+import ScrollToTop from './shared/ScrollToTop'
+import Profile from './main/Profile'
+import MyShows from './main/MyShows'
+import IndividualShow from './main/IndividualShow'
 import Modal from 'react-modal'
-import Login from './Login'
-import Signup from './Signup'
+import Login from './auth/Login'
+import Signup from './auth/Signup'
 require('dotenv').config()
 
 Modal.setAppElement('#root') //Necessary for settings modal
