+++ front-end/src/auth/LoginButton.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import './css/LoginButton.css'
+import '../css/LoginButton.css'
 
 function LoginButton() {
   return (
