+++ front-end/src/auth/LogoutButton.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useHistory } from 'react-router-dom'
-import { AuthContext } from './App.js'
-import './css/LogoutButton.css'
+import { AuthContext } from '../App.js'
+import '../css/LogoutButton.css'
 
 function LogoutButton() {
   const { setLoggedInUser } = React.useContext(AuthContext)
