+++ front-end/src/auth/SignupButton.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import './css/SignupButton.css'
+import '../css/SignupButton.css'
 
 function SignupButton() {
   return (
