+++ front-end/src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './css/index.css'
 import App from './App'
-import reportWebVitals from './reportWebVitals'
+import reportWebVitals from './shared/reportWebVitals'
 
 ReactDOM.render(
   <React.StrictMode>
