+++ front-end/src/main/MyShows.js
@@ -1,21 +1,21 @@
 import React, { useEffect, useState } from 'react'
 import AsyncSelect from 'react-select/async'
-import Header from './Header'
-import Footer from './Footer'
+import Header from '../shared/Header'
+import Footer from '../shared/Footer'
 import axios from 'axios'
-import { mockShowAPI, mockShowImage } from './MockData'
-import './css/MyShows.css'
+import { mockShowAPI, mockShowImage } from '../shared/MockData'
+import '../css/MyShows.css'
 import Modal from 'react-modal'
 import Select from 'react-select'
 import { Link, useHistory } from 'react-router-dom'
-import amazon from './Images/amazon.jpg'
-import crunchyroll from './Images/crunchyroll.png'
-import disney from './Images/disney.jpg'
-import hbo from './Images/hbo.jpg'
-import hulu from './Images/hulu.png'
-import netflix from './Images/netflix.png'
-import other from './Images/other.png'
-import { platforms, textToValue } from './DropdownOptions'
+import amazon from '../Images/amazon.jpg'
+import crunchyroll from '../Images/crunchyroll.png'
+import disney from '../Images/disney.jpg'
+import hbo from '../Images/hbo.jpg'
+import hulu from '../Images/hulu.png'
+import netflix from '../Images/netflix.png'
+import other from '../Images/other.png'
+import { platforms, textToValue } from '../shared/DropdownOptions'
 require('dotenv').config()
 
 const ShowGrid = (props) => {
