+++ front-end/src/main/Profile.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react'
-import Footer from './Footer'
+import Footer from '../shared/Footer'
 import axios from 'axios'
-import './css/Profile.css'
-import Header from './Header'
+import '../css/Profile.css'
+import Header from '../shared/Header'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
 import {
   createMockUser,
@@ -10,10 +10,10 @@ import {
   mockUserImage,
   mockShowImage,
   mockUserUpdate
-} from './MockData'
+} from '../shared/MockData'
 import { Link } from 'react-router-dom'
 import Modal from 'react-modal'
-import { AuthContext } from './App'
+import { AuthContext } from '../App'
 require('dotenv').config()
 
 /* UserInfo displays username, user bio, and user profile picture */
