+++ front-end/src/shared/Footer.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import './css/Footer.css'
+import '../css/Footer.css'
 import { Link } from 'react-router-dom'
-import trakt from './Images/trakt.png'
+import trakt from '../Images/trakt.png'
 
 function Footer() {
   return (
